修正介面顯示
*修改 Donut 圖 Used Percentage 與 Used Text 字體大小
*Usage Card 下方的已使用與可使用的數值，換算的單位從1000改為1024
*Usage Card 下方的已使用與可使用的單修改為大寫

+++ app/src/main/java/com/cephmonitor/cephmonitor/layout/dialog/fixed/LoginLanguageDialog.java
@@ -137,9 +137,9 @@ public class LoginLanguageDialog extends SettingDialog {
     public void show() {
         SettingStorage settingStorage = new SettingStorage(getContext());
         selectedId = settingStorage.getLanguage();
-        english.filedValue.setState(english.getTag() == selectedId);
-        chinese.filedValue.setState(chinese.getTag() == selectedId);
-        japanese.filedValue.setState(japanese.getTag() == selectedId);
+        english.filedValue.setState(((int) english.getTag()) == selectedId);
+        chinese.filedValue.setState(((int) chinese.getTag()) == selectedId);
+        japanese.filedValue.setState(((int) japanese.getTag()) == selectedId);
         setTitle(getContext().getString(R.string.settings_profile_language));
         super.show();
     }
